+++ components/logins/src/schema.rs
@@ -194,6 +194,7 @@ pub(crate) static LAST_SYNC_META_KEY: &str = "last_sync_time";
 pub(crate) static GLOBAL_STATE_META_KEY: &str = "global_state_v2";
 pub(crate) static GLOBAL_SYNCID_META_KEY: &str = "global_sync_id";
 pub(crate) static COLLECTION_SYNCID_META_KEY: &str = "passwords_sync_id";
+pub(crate) static CHECKPOINT_KEY: &str = "checkpoint";
 
 pub(crate) fn init(db: &Connection) -> Result<()> {
     let user_version = db.query_one::<i64>("PRAGMA user_version")?;